Add validation to api methods

+++ application/api/employee.1/deleteEmployeeById.js
@@ -1,4 +1,16 @@
-async ({ employeeid }) => {
-  await domain.db.delete('Employee', { employeeid });
-  return { result: 'success' };
-};
+({
+  caption: 'Delete employee by id',
+  description: 'Delete employee record frow database',
+  access: 'logged',
+  parameters: {
+    employeeid: 'number',
+  },
+  method: async ({ employeeid }) => {
+    if (!employeeid) {
+      return new Error('No id provided');
+    }
+
+    await domain.db.delete('Employee', { employeeid });
+    return { result: 'success' };
+  },
+});
